[FLINK-29501] Add option to override job vertex parallelisms during job submission (#21023)
This allows to change the job vertex parallelism of a JobGraph during job submission time without
having to modify the JobGraph upfront. This is intended to be used for (auto)scaling Flink jobs
based on metric observations.

The initial idea was to add a new field to the payload of the job submit Rest endpoint. However, it
is probably more practical to handle the overrides in the same way as other PipelineOptions already
do it, i.e. via the configuration.

The implementation is deliberately lenient with respect to the presence of job vertices. If vertices
have been removed or new ones have been added, only the ones found will have their parallelism
overrides. The verification should be performed by the caller, not by Flink. In particular, we want
to support scenarios where users modify the deployment and we might not yet have overrides for all
vertices.